refactor: extract Race class to separate file
